requirements.txt
fastapi
uvicorn
watchfiles
python-dotenv

langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
google-genai

pypdf
faiss-cpu
sentence-transformers
torch
gradio
